# [11.x] Optimize boostrap time by using hashtable to store providers