It should slightly optimize bootstrap time, especially for large applications with many providers. I've encountered that the method getProvider has O(n) complexity, but it could be optimized to O(1).

- Application::register
- Application::getProvider

<img width="397" alt="Screenshot 2024-05-08 at 15 30 05" src="https://github.com/laravel/framework/assets/6762480/5fe32687-287f-478e-89a6-5a049fd1dcb8">

Of course, this time from Blackfire is not true, and the original value is smaller, but still executing that function ~63k times isn't a good thing. 

Seems like serviceProviders hashtable is enough, and we can get rid of loadedProviders.