@browner12 2024-05-08
Maybe I'm not understanding something here, but isn't the point of having separate "registered" and "loaded" service providers so we can defer loading of ones that we won't necessarily need on every request? Does this change maintain that functionality?