@sarven 2024-05-08
@browner12 if you mean deferred providers, those changes don't touch that functionality directly. Deferred providers are in the property: deferredServices

loadProviders property was used only in the method: markAsRegistered, where a provider is added to serviceProviders, so it looks like using only one array as the source of truth should be ok.