@taylorotwell 2024-05-08
Can you share benchmarks of speed improvements on a blank Laravel application that makes a single database query?