@bert-w 2024-05-08
Hasn't logic changed now? Because `getProvider()` used to check for `instanceof` (in `getProviders()`) and now it doesnt anymore.