@sarven 2024-05-09
@taylorotwell A fresh application contains a small number of providers, so it isn't even shown by Blackfire as a significant cost. However, I've added 300 providers to some fresh app, and here are the results:

**BEFORE**

<img width="1426" alt="Screenshot 2024-05-09 at 15 37 30" src="https://github.com/laravel/framework/assets/6762480/06223804-c15d-4188-9da9-3972311bc10d">
<img width="397" alt="Screenshot 2024-05-09 at 15 37 59" src="https://github.com/laravel/framework/assets/6762480/a67a04f9-873f-4702-9e20-d496d36039dc">


**AFTER**
<img width="1517" alt="Screenshot 2024-05-09 at 15 39 38" src="https://github.com/laravel/framework/assets/6762480/fa86aaf8-4baf-4df8-9a5a-83c1b3d0f184">
<img width="430" alt="Screenshot 2024-05-09 at 15 39 48" src="https://github.com/laravel/framework/assets/6762480/c52aae1c-d0bc-4df4-8d82-8ffc58d7d65c">

After my change, Blackfire doesn't show that method, because the cost is insignificant. There are **((n - 1) * n) / 2** of executions, so for small apps, it isn't important, but for bigger applications with more providers like 100, 200, 300, or even more it should be a nice optimization having an impact on every request.