@sarven 2024-05-13
@driesvints thanks for clarifying 

@taylorotwell could you have another look? I'm not sure if you received a notification about my response, because it was written when the PR was still marked as a draft.