@taylorotwell 2024-05-17
Sorry, maybe I'm missing it, but how long does it take to run an entire Laravel request (in milliseconds) for a normal app with a single database query before and after this PR?