@sarven 2024-05-18
@taylorotwell As I've written, this isn't relevant for small apps with only a few providers. So I can't prepare such a comparison for "a normal app with one database query", because it isn't possible to show any optimization in that case. It would be useful only for larger apps with many providers. I've seen some monolithic apps using Laravel with over 300 providers, so this isn't unusual. Here is a comparison of an app with 300 providers:

`ab -n 1000 -c 1 http://127.0.0.1:8080/`

**BEFORE**
```
Server Software:        nginx/1.25.5
Server Hostname:        127.0.0.1
Server Port:            8080

Document Path:          /
Document Length:        20 bytes

Concurrency Level:      1
Time taken for tests:   24.429 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      1146000 bytes
HTML transferred:       20000 bytes
Requests per second:    40.93 [#/sec] (mean)
Time per request:       24.429 [ms] (mean)
Time per request:       24.429 [ms] (mean, across all concurrent requests)
Transfer rate:          45.81 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.5      0       5
Processing:    17   24   8.3     22      95
Waiting:       17   24   8.3     22      95
Total:         17   24   8.3     22      95

Percentage of the requests served within a certain time (ms)
  50%     22
  66%     24
  75%     24
  80%     25
  90%     27
  95%     32
  98%     65
  99%     70
 100%     95 (longest request)
```

**AFTER**
```
Concurrency Level:      1
Time taken for tests:   20.777 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      1146000 bytes
HTML transferred:       20000 bytes
Requests per second:    48.13 [#/sec] (mean)
Time per request:       20.777 [ms] (mean)
Time per request:       20.777 [ms] (mean, across all concurrent requests)
Transfer rate:          53.86 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       4
Processing:    14   20   6.7     19      83
Waiting:       14   20   6.7     18      82
Total:         15   21   6.7     19      83

Percentage of the requests served within a certain time (ms)
  50%     19
  66%     20
  75%     21
  80%     22
  90%     24
  95%     32
  98%     44
  99%     59
 100%     83 (longest request)
```

BTW It is hard to measure fairly, because the difference is quite small 2-3ms per request, so the best way to present that optimization is to look at the profiler (screens in my previous comment).