@taylorotwell 2024-05-20
@sarven sorry to drag this PR out a bit, but is it at all possible to make the relevant performance improvements with minimal others changes? For example while keeping the `loadedProviders` property?