@sarven 2024-05-21
@taylorotwell no problem, changed. I thought also about creating a new private method like getProviderByClassName to keep the behavior of the current getProvider the same. Currently, for example, it's possible to get the first provider of the base class like ServiceProvider, because there is used instanceof, but it doesn't seem like something useful and I assume that no one uses that method in that way, so changing the behavior a little bit shouldn't be a problem. What do you think?